// x-pack/solutions/observability/plugins/observability_agent_builder/server/data_registry/data_registry_types.d.ts
export interface TraceMetrics {
    latencyUs: number | null;
    throughputPerMin: number | null;
    errorRate: number | null;
}
export type ApmConnectionStatsEntry = {
    type: 'service';
    serviceName: string;
    metrics: TraceMetrics;
} | {
    type: 'dependency';
    dependencyName: string;
    spanType: string;
    spanSubtype: string;
    metrics: TraceMetrics;
};